+++ scripts/macros/2014/classFeatures/illrigger/combatMasteries/lies.js
@@ -1,6 +1,10 @@
 import {constants, genericUtils, itemUtils, workflowUtils} from '../../../../../utils.js';
-async function attack({trigger, workflow}) {
-    if (!workflow.activity || !constants.meleeWeaponTypes.includes(workflow.item?.system?.type?.value)) return;
+async function attack({trigger: {entity: item}, workflow}) {
+    if (!workflow.activity || !workflow.item) return;
+    let baseItem = workflow.item.system.type?.baseItem;
+    if (!baseItem) return;
+    let weaponType = itemUtils.getConfig(item, 'weaponType');
+    if (baseItem != weaponType) return;
     await workflowUtils.swapAttackAbility(workflow, 'cha');
 }
 export let combatMasteryLies = {
@@ -15,5 +19,15 @@ export let combatMasteryLies = {
                 priority: 25
             }
         ]
-    }
+    },
+    config: [
+        {
+            value: 'weaponType',
+            label: 'CHRISPREMADES.Config.WeaponType',
+            type: 'select',
+            default: 'battleaxe',
+            category: 'mechanics',
+            options: constants.getBaseMeleeWeaponOptions
+        }
+    ]
 };