+++ scripts/macros/2024/classFeatures/ranger/gloomStalker/umbralSight.js
@@ -1,0 +1,6 @@
+import {umbralSight as umbralSightLegacy} from '../../../../../legacyMacros.js';
+export let umbralSight = {
+    ...umbralSightLegacy,
+    version: '1.3.81',
+    rules: 'modern',
+};