+++ scripts/macros/2024/feats/slasher.js
@@ -1,6 +1,6 @@
 import {slasher as slasherLegacy} from '../../../legacyMacros.js';
 export let slasher = {
     ...slasherLegacy,
-    version: '1.2.36',
+    version: '1.3.81',
     rules: 'modern'
 };